User profile action, don't allow current user to be deleted.

+++ app/Http/Controllers/API/UserController.php
@@ -116,6 +116,10 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        $currentUser = \Auth::user();
+        if($currentUser->id == $id) {
+            abort(403, 'Unauthorized action.');
+        }
         $user = User::find($id);
         $user->delete();
         return $user;
